[libs] E: Move Binding for posix_fadvise()
This commits moves the binding for the posix_fadvise() system call to
the syscall crate.
