+++ src/libs/posix/src/fcntl/mod.rs
@@ -21,58 +21,6 @@ use ::syscall::fcntl;
 ///
 /// # Description
 ///
-/// Provides advice about the use of a file descriptor.
-///
-/// # Parameters
-///
-/// - `fd`: File descriptor.
-/// - `offset`: Offset in bytes.
-/// - `len`: Length in bytes.
-/// - `advice`: Advice to provide.
-///
-/// # Returns
-///
-/// Upon success, `posix_fadvise()` returns zero. Otherwise, it `-1` and sets `errno` to indicate
-/// the error.
-///
-/// # Safety
-///
-/// This function is unsafe because it may access global variables.
-///
-/// It is safe to call this function if the following conditions are met:
-/// - This function is not called from multiple threads at the same time.
-///
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn posix_fadvise(fd: c_int, offset: i64, len: i64, advice: c_int) -> c_int {
-    ::syslog::trace!(
-        "posix_fadvise(): fd={:?}, offset={:?}, len={:?}, advice={:?}",
-        fd,
-        offset,
-        len,
-        advice
-    );
-
-    // Run system call and check for errors.
-    match fcntl::posix_fadvise(fd, offset, len, advice) {
-        Ok(()) => 0,
-        Err(error) => {
-            ::syslog::error!(
-                "posix_fadvise(): failed (fd={:?}, offset={:?}, len={:?}, advice={:?}, error={:?})",
-                fd,
-                offset,
-                len,
-                advice,
-                error
-            );
-            *__errno_location() = error.code.get();
-            -1
-        },
-    }
-}
-
-///
-/// # Description
-///
 /// Renames a file relative to a directory file descriptor.
 ///
 /// # Parameters
