+++ src/libs/syscall/src/fcntl/bindings/mod.rs
@@ -7,4 +7,5 @@
 
 pub mod fcntl;
 pub mod open;
+pub mod posix_fadvise;
 pub mod posix_fallocate;
